chore: rename writer module
