+++ src/ffi/mod.rs
@@ -1,4 +1,4 @@
 pub mod utils;
 pub mod reader;
-pub mod writers;
+pub mod writer;
 pub mod memory;
