+++ src/ffi/writer.rs
